A little experiment on top of #256. If you have [Node](https://nodejs.org/) and [`wasm-tools`](https://github.com/bytecodealliance/wasm-tools) installed, you can run it like this:

```sh
./gradbench run --eval './gradbench repo eval kmeans' --tool 'wasm-tools parse -o tools/floretta/kmeans.wa{sm,t} && node tools/floretta/run.mjs'
```

You can also try using [`wasm-opt`](https://github.com/WebAssembly/binaryen) to inline all the WebAssembly into one function:

```sh
./gradbench run --eval './gradbench repo eval kmeans' --tool 'wasm-tools parse -o tools/floretta/kmeans.wa{sm,t} && npx --package binaryen wasm-opt -O3 -o tools/floretta/kmeans{-opt,}.wasm && node tools/floretta/run.mjs --opt'
```

Here's a similar command trying `wasm-opt` with the Wasmtime engine instead of Node:

```sh
./gradbench run --eval './gradbench repo eval kmeans' --tool 'wasm-tools parse -o tools/floretta/kmeans.wa{sm,t} && npx --package binaryen wasm-opt -O3 -o tools/floretta/kmeans{-opt,}.wasm && cargo run --package gradbench-floretta --release -- --opt'
```

And for completeness, here's the original command using Wasmtime without `wasm-opt`:

```sh
./gradbench run --eval './gradbench repo eval kmeans' --tool 'cargo run --package gradbench-floretta --release'
```

I see some difference in performance between Node and Wasmtime, but, surprisingly, I see a _much bigger_ performance difference when using `wasm-opt` versus not. For instance, with the `k=1000,n=10000,d=16` workload, here are the numbers I get on my M1 MacBook:

- Wasmtime ~~without `wasm-opt`~~: 562ms evaluate
- Node ~~without `wasm-opt`~~: 364ms evaluate
- Wasmtime **with `wasm-opt`**: 116ms evaluate
- Node **with `wasm-opt`**: 89ms evaluate

This is surprising to me because, if you look at the output from `wasm-opt`, you can see that all it did is inline all the functions into one:

```sh
wasm-tools print -f tools/floretta/kmeans-opt.wasm
```

I guess I'm just surprised that Node and Wasmtime don't do inlining on their own.